+++ Repopulator/Matchers/TagMatcher.cs
@@ -14,7 +14,7 @@ namespace Repopulator.Matchers
     internal class TagMatcher : RepopulatorMatcher
     {
         private string[] _fileList;
-        private int _currentFile = 0;
+        private int _currentFile;
 
         private CsvToDicomColumn _indexer;
         private DicomTag _indexerTag;
@@ -33,7 +33,7 @@ namespace Repopulator.Matchers
             if(_indexer == null)
                 throw new ArgumentException("No valid indexer could be found, there must be a column in the map for either SOP, Series or Study instance UIDs");
 
-            using (var reader = new CsvReader(map.CsvFile.OpenText()))
+            using (var reader = new CsvReader(map.CsvFile.OpenText(), System.Globalization.CultureInfo.CurrentCulture))
             {
                 reader.Configuration.TrimOptions = TrimOptions.Trim;
 
@@ -77,13 +77,14 @@ namespace Repopulator.Matchers
 
             try
             {
-                var df = DicomFile.Open(_fileList[_currentFile]);
+                var currentFile = new FileInfo(_fileList[_currentFile]);
+                var df = DicomFile.Open(currentFile.FullName);
                 var seek = df.Dataset.GetValue<string>(_indexerTag, 0);
 
                 if(!_indexerToRowMap.ContainsKey(seek))
                     throw new Exception($"Csv did not contain a value for {_indexerTag} {seek} which was found in file '{_fileList[_currentFile]}'");
 
-                return new RepopulatorJob(Map,df,_indexerToRowMap[seek]);
+                return new RepopulatorJob(Map,currentFile,df,_indexerToRowMap[seek]);
             }
             finally
             {