+++ Tests/DicomRepopulatorTests.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
+using System.Threading;
 using Dicom;
 using NUnit.Framework;
 using Repopulator;
@@ -32,8 +32,9 @@ namespace Tests
             if (Directory.Exists(_seriesFilesBase)) Directory.Delete(_seriesFilesBase, true);
         }
 
-        [Test]
-        public void Test_AbsolutePath_InCsv()
+        [TestCase(true)]
+        [TestCase(false)]
+        public void Test_AbsolutePath_InCsv(bool deleteAsYouGo)
         {
             //Create a dicom file in the input dir /subdir/IM_0001_0013.dcm
             var inputDicom = CreateInputFile(TestData.IMG_013,
@@ -50,11 +51,18 @@ namespace Tests
                 inputCsv,
                 null,
                 inputDicom.Directory.Parent,
-                (o) => o.FileNameColumn = "File");
+                o =>
+                {
+                    o.FileNameColumn = "File";
+                    o.DeleteAsYouGo = deleteAsYouGo;
+                }
+                );
             
             //anonymous image should appear in the subdirectory of the out dir
             var expectedOutFile = new FileInfo(Path.Combine(outDir.FullName, "subdir", nameof(TestData.IMG_013) + ".dcm"));
             FileAssert.Exists(expectedOutFile);
+
+            Assert.AreEqual(!deleteAsYouGo,File.Exists(inputDicom.FullName));
         }
 
         [TestCase("./subdir/IM-0001-0013.dcm")]
@@ -76,7 +84,7 @@ namespace Tests
                 inputCsv,
                 null,
                 inputDicom.Directory.Parent,
-                (o) => o.FileNameColumn = "File");
+                o => o.FileNameColumn = "File");
             
             //anonymous image should appear in the subdirectory of the out dir
             var expectedOutFile = new FileInfo(Path.Combine(outDir.FullName, "subdir", Path.GetFileName(TestData.IMG_013)));
@@ -134,7 +142,7 @@ namespace Tests
                 CreateExtraMappingsFile("ID:PatientID"), inFile.Directory,
                 
                 //Give it BasicTest.csv 
-                (o) => o.InputCsv= Path.Combine(TestContext.CurrentContext.TestDirectory, "BasicTest.csv"));
+                o => o.InputCsv= Path.Combine(TestContext.CurrentContext.TestDirectory, "BasicTest.csv"));
 
             //Anonymous dicom image should exist
             var expectedFile = new FileInfo(Path.Combine(outDir.FullName, nameof(TestData.IMG_013) + ".dcm"));
@@ -168,7 +176,7 @@ namespace Tests
                 NumThreads = 4
             };
 
-            int result = new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory).Process(options);
+            int result = new DicomRepopulatorProcessor().Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
 
             Assert.True(File.Exists(expectedFile), "Expected output file {0} to exist", expectedFile);
@@ -198,7 +206,7 @@ namespace Tests
                 NumThreads = 4
             };
 
-            int result = new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory).Process(options);
+            int result = new DicomRepopulatorProcessor().Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
 
             Assert.True(File.Exists(expectedFile), "Expected output file {0} to exist", expectedFile);
@@ -230,7 +238,7 @@ namespace Tests
                 NumThreads = 1
             };
 
-            int result = new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory).Process(options);
+            int result = new DicomRepopulatorProcessor().Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
 
             Assert.True(File.Exists(expectedFile), "Expected output file {0} to exist", expectedFile);
@@ -262,7 +270,7 @@ namespace Tests
                 NumThreads = 1
             };
 
-            int result = new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory).Process(options);
+            int result = new DicomRepopulatorProcessor().Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
 
             Assert.True(File.Exists(expectedFile), "Expected output file {0} to exist", expectedFile);
@@ -295,7 +303,7 @@ namespace Tests
                 NumThreads = 1
             };
 
-            int result = new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory).Process(options);
+            int result = new DicomRepopulatorProcessor().Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
 
             Assert.True(File.Exists(expectedFile1), "Expected output file {0} to exist", expectedFile1);
@@ -339,7 +347,7 @@ namespace Tests
                 NumThreads = 4
             };
 
-            int result = new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory).Process(options);
+            int result = new DicomRepopulatorProcessor().Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
 
             Assert.True(File.Exists(expectedFile1), "Expected output file {0} to exist", expectedFile1);
@@ -387,9 +395,9 @@ namespace Tests
 
         /// <summary>
         /// Creates a new input directory with the name of the calling method (under <see cref="_inputFileBase"/>) then creates
-        /// the given dicom file (<paramref name="bytes"/>) at <paramref name="filename"/> location (which can include subdirectories)
+        /// the given dicom file (<paramref name="testFile"/>) at <paramref name="filename"/> location (which can include subdirectories)
         /// </summary>
-        /// <param name="bytes">The dicom files raw bytes</param>
+        /// <param name="testFile">The dicom file</param>
         /// <param name="filename">The filename to write out e.g. "my.dcm" or "mysubdir/my.dcm"</param>
         /// <param name="memberName"></param>
         /// <returns></returns>
@@ -400,7 +408,7 @@ namespace Tests
             Directory.CreateDirectory(inputDirPath);
             var toReturn = new FileInfo(Path.Combine(inputDirPath, filename));
             
-            toReturn.Directory.Create();
+            toReturn.Directory?.Create();
 
             return TestData.Create(toReturn, testFile);
         }
@@ -409,9 +417,11 @@ namespace Tests
         /// are no errors during the run and th
         /// </summary>
         /// <param name="expectedDone"></param>
+        /// <param name="expectedErrors"></param>
         /// <param name="inputCsv"></param>
         /// <param name="inputExtraMapping"></param>
         /// <param name="inputDicomDirectory"></param>
+        /// <param name="adjustOptions"></param>
         /// <param name="memberName"></param>
         /// <returns></returns>
         private DirectoryInfo AssertRunsSuccesfully(int expectedDone, int expectedErrors, FileInfo inputCsv, FileInfo inputExtraMapping,
@@ -428,7 +438,7 @@ namespace Tests
                 out DicomRepopulatorOptions options,
                 out DirectoryInfo outputDirPath);
             
-            int result = processor.Process(options);
+            int result = processor.Process(options,CancellationToken.None);
             Assert.AreEqual(0, result);
             
             foreach (var log in processor.MemoryLogTarget.Logs)
@@ -465,7 +475,7 @@ namespace Tests
 
             adjustOptions?.Invoke(options);
 
-            return new DicomRepopulatorProcessor(TestContext.CurrentContext.TestDirectory);
+            return new DicomRepopulatorProcessor();
         }
 
     }