+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added logging to ProcessJob for outPath in DicomRepopulatorProcessor
 
 ...
+- Dependency updates
+- LGTM fixes
+- Report if error encountered when saving or loading a file
+
+### Added
+
+- DeleteAsYouGo mode where input files are deleted after processing to the output directory
 
 ## [1.1.0] - 2020-02-07